+++ Electronics_Shop_17.Model/Requests/UpdateProduct.cs
@@ -13,7 +13,7 @@ namespace Electronics_Shop_17.Model.Requests
         public string? Brand { get; set; }
         public string? Model { get; set; }
         public string? Description { get; set; }
-        public decimal? Price { get; set; }
+        public double? Price { get; set; }
         public int? AllColorsStock { get; set; }
 
         public int? ProductCategoryId { get; set; }
