+++ Electronics_Shop_17.Model/SearchObjects/SearchOrderItem.cs
@@ -10,8 +10,8 @@ namespace Electronics_Shop_17.Model.SearchObjects
     {
         public int? Id { get; set; }
         public int? Quantity { get; set; }
-        public decimal? Price { get; set; }
-        public decimal? FinalPrice { get; set; }
+        public double? Price { get; set; }
+        public double? FinalPrice { get; set; }
 
         public int? OrderId { get; set; }
     }
