+++ Electronics_Shop_17.Model/SearchObjects/SearchProduct.cs
@@ -13,8 +13,8 @@ namespace Electronics_Shop_17.Model.SearchObjects
         public string? Brand { get; set; }
         public string? Model { get; set; }
         public string? Description { get; set; }
-        public decimal? PriceLow { get; set; }
-        public decimal? PriceHigh { get; set; }
+        public double? PriceLow { get; set; }
+        public double? PriceHigh { get; set; }
         public int? AllColorsStock { get; set; }
 
         public int? ProductCategoryId { get; set; }
