+++ Electronics_Shop_17.Services/Database/OrderItem.cs
@@ -10,12 +10,12 @@ namespace Electronics_Shop_17.Services.Database
     {
         public int Id { get; set; }
         public int Quantity { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
 
         public int OrderId { get; set; }
         public int ProductId { get; set; }
         public virtual Product Product { get; set; }
-        public decimal FinalPrice { get; set; }
+        public double FinalPrice { get; set; }
         public int ProductColorId { get; set; }
         public virtual ProductColor ProductColor { get; set; }
     }
